print_string: Rename MemPrint to PrintString; namespace mem_print to print_string
