+++ tests/PrintStringTest/Makefile
@@ -1,6 +1,6 @@
 # See https://github.com/bxparks/UnixHostDuino for documentation about this
 # Makefile to compile and run Arduino programs natively on Linux or MacOS.
 
-APP_NAME := MemPrintTest
+APP_NAME := PrintStringTest
 ARDUINO_LIBS := AUnit AceUtils
 include ../../../UnixHostDuino/UnixHostDuino.mk
